+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but MINOR version are breaking before 1.0.0
+
 
 ## [Unreleased]
 ### Added
@@ -12,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - 
 
 ## [0.2.0] - 2024-10-15
-### Minor interface breaking chages:
+### Interface breaking chages:
 because: for many functions, the return type is wayyy to high compared to what is possible for it return
          also usize isn't really meant to be an arguement type, so I'm changing that also to u16 as soon as possible