+++ CHANGELOG.md
@@ -17,58 +17,109 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 because: for many functions, the return type is wayyy to high compared to what is possible for it return
          also usize isn't really meant to be an arguement type, so I'm changing that also to u16 as soon as possible
 
--   ff2n(n: usize) -> Result<Array2<i32>, String>
+-   `ff2n(n: usize) -> Result<Array2<i32>, String>`
+
     changed to 
+
+    ```rust
     ff2n(n: u16) Result<Array2<i16>, String>
+    ```
 
--   gsd(levels: &[u16], reduction: usize, n_designs: usize) -> Result<Vec<Array2<u16>>, String>
+-   `gsd(levels: &[u16], reduction: usize, n_designs: usize) -> Result<Vec<Array2<u16>>, String>`
+
     changed to
+
+    ```rust
     gsd(levels: &[u16], reduction: u16, n_designs: u16) -> Result<Vec<Array2<u16>>, String>
+    ```
+    
+-   `bbdesign(n: usize) -> Result<Array2<i32>, String>`
 
--   bbdesign(n: usize) -> Result<Array2<i32>, String> 
     changed to 
+
+    ```rust
     bbdesign(n: u16) -> Result<Array2<i16>, String> 
+    ```
 
--   bbdesign_center(n: usize, center: usize) -> Result<Array2<i32>, String>
+-   `bbdesign_center(n: usize, center: usize) -> Result<Array2<i32>, String>`
+
     changed to 
+
+    ```rust
     bbdesign_center(n: u16, center: u16) -> Result<Array2<i16>, String>
+    ```
 
--   ccdesign(n: usize, center: &[u32], alpha: Alpha, face: Face) -> Result<Array2<f32>, String> 
+-   `ccdesign(n: usize, center: &[u32], alpha: Alpha, face: Face) -> Result<Array2<f32>, String>`
+
     changed to 
+
+    ```rust
     ccdesign(n: u16, center: &[u32], alpha: Alpha, face: Face) -> Result<Array2<f32>, String> 
+    ```
 
--   star(n: usize, alpha: Alpha, center: &[u32]) -> (Array2<f32>, f32)
+-   `star(n: usize, alpha: Alpha, center: &[u32]) -> (Array2<f32>, f32)`
+
     changed to
+
+    ```rust
     star(n: u16, alpha: Alpha, center: &[u32]) -> (Array2<f32>, f32)
+    ```
 
--   pbdesign(n: u32) -> Array2<i32>
+-   `pbdesign(n: u32) -> Array2<i32>`
+
     changed to
+
+    ```rust
     pbdesign(n: u32) -> Array2<i16>
+    ```
 
--   fracfact(design: &str) -> Array2<i32>
+-   `fracfact(design: &str) -> Array2<i32>`
+
     changed to
+
+    ```rust
     fracfact(design: &str) -> Array2<i16>
+    ```
 
--   lhs_classic(n: usize, samples: usize, random_state: u64) -> Array2<f32>
+-   `lhs_classic(n: usize, samples: usize, random_state: u64) -> Array2<f32>`
+
     changed to
+
+    ```rust
     lhs_classic(n: u16, samples: u16, random_state: u64) -> Array2<f32>
+    ```
 
--   lhs_centered(n: usize, samples: usize, random_state: u64) -> Array2<f32>
+-   `lhs_centered(n: usize, samples: usize, random_state: u64) -> Array2<f32>`
+
     changed to
+
+    ```rust
     lhs_centered(n: u16, samples: u16, random_state: u64) -> Array2<f32>
+    ```
 
--   lhs_maximin(n: usize, samples: usize, random_state: u64, iterations: u16) -> Array2<f32>
+-   `lhs_maximin(n: usize, samples: usize, random_state: u64, iterations: u16) -> Array2<f32>`
+
     changed to
+
+    ```rust
     lhs_maximin(n: u16, samples: u16, random_state: u64, iterations: u16) -> Array2<f32>
+    ```
 
--   lhs_correlate(n: usize, samples: usize, random_state: u64, iterations: u16) -> Array2<f32>
+-   `lhs_correlate(n: usize, samples: usize, random_state: u64, iterations: u16) -> Array2<f32>`
+
     changed to
+
+    ```rust
     lhs_correlate(n: u16, samples: u16, random_state: u64, iterations: u16) -> Array2<f32>
+    ```
 
--   lhs_mu(n: usize, samples: usize, random_state: u64) -> Array2<f32>
+-   `lhs_mu(n: usize, samples: usize, random_state: u64) -> Array2<f32>`
+
     changed to 
-    lhs_mu(n: u16, samples: u16, random_state: u64) -> Array2<f32>
 
+    ```rust
+    lhs_mu(n: u16, samples: u16, random_state: u64) -> Array2<f32>
+    ```
 
 ### Other changes:
 -   fixing documentation