+++ CHANGELOG.md
@@ -14,8 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.2.0] - 2024-10-15
 ### Interface breaking chages:
-because: for many functions, the return type is wayyy to high compared to what is possible for it return
-         also usize isn't really meant to be an arguement type, so I'm changing that also to u16 as soon as possible
+because: for many functions, the return type is way too high compared to what is possible for it return
+also usize isn't really meant to be an argument type, so I'm changing that to u16 
 
 -   `ff2n(n: usize) -> Result<Array2<i32>, String>`